+++ components/ads-client/src/client.rs
@@ -7,7 +7,7 @@ use std::collections::HashMap;
 use std::time::Duration;
 
 use crate::http_cache::{ByteSize, CachePolicy, HttpCache};
-use crate::mars::ad_request::AdPlacementRequest;
+use crate::mars::ad_request::{AdPlacementRequest, AdRequestFlags};
 use crate::mars::ad_response::{AdImage, AdResponse, AdResponseValue, AdSpoc, AdTile};
 use crate::mars::error::{RecordClickError, RecordImpressionError, ReportAdError};
 use crate::mars::{MARSClient, ReportReason};
@@ -160,11 +160,12 @@ where
     pub fn request_image_ads(
         &self,
         ad_placement_requests: Vec<AdPlacementRequest>,
+        flags: AdRequestFlags,
         options: Option<CachePolicy>,
         ohttp: bool,
     ) -> Result<HashMap<String, AdImage>, RequestAdsError> {
         let response = self
-            .request_ads::<AdImage>(ad_placement_requests, options, ohttp)
+            .request_ads::<AdImage>(ad_placement_requests, flags, options, ohttp)
             .inspect_err(|e| {
                 self.telemetry.record(e);
             })?;
@@ -175,10 +176,11 @@ where
     pub fn request_spoc_ads(
         &self,
         ad_placement_requests: Vec<AdPlacementRequest>,
+        flags: AdRequestFlags,
         options: Option<CachePolicy>,
         ohttp: bool,
     ) -> Result<HashMap<String, Vec<AdSpoc>>, RequestAdsError> {
-        let result = self.request_ads::<AdSpoc>(ad_placement_requests, options, ohttp);
+        let result = self.request_ads::<AdSpoc>(ad_placement_requests, flags, options, ohttp);
         result
             .inspect_err(|e| {
                 self.telemetry.record(e);
@@ -192,10 +194,11 @@ where
     pub fn request_tile_ads(
         &self,
         ad_placement_requests: Vec<AdPlacementRequest>,
+        flags: AdRequestFlags,
         options: Option<CachePolicy>,
         ohttp: bool,
     ) -> Result<HashMap<String, AdTile>, RequestAdsError> {
-        let result = self.request_ads::<AdTile>(ad_placement_requests, options, ohttp);
+        let result = self.request_ads::<AdTile>(ad_placement_requests, flags, options, ohttp);
         result
             .inspect_err(|e| {
                 self.telemetry.record(e);
@@ -209,6 +212,7 @@ where
     fn request_ads<A>(
         &self,
         placements: Vec<AdPlacementRequest>,
+        flags: AdRequestFlags,
         options: Option<CachePolicy>,
         ohttp: bool,
     ) -> Result<AdResponse<A>, RequestAdsError>
@@ -219,7 +223,7 @@ where
         let cache_policy = options.unwrap_or_default();
         let (mut response, request_hash) =
             self.client
-                .fetch_ads::<A>(context_id, placements, cache_policy, ohttp)?;
+                .fetch_ads::<A>(context_id, flags, placements, cache_policy, ohttp)?;
         response.enrich_callbacks(&request_hash);
         Ok(response)
     }
@@ -289,7 +293,12 @@ mod tests {
         let mars_client = MARSClient::new(Environment::Test, None, MozAdsTelemetryWrapper::noop());
         let ads_client = new_with_mars_client(mars_client);
 
-        let result = ads_client.request_image_ads(make_happy_placement_requests(), None, false);
+        let result = ads_client.request_image_ads(
+            make_happy_placement_requests(),
+            AdRequestFlags::default(),
+            None,
+            false,
+        );
         assert!(result.is_ok());
         m.assert();
     }
@@ -308,7 +317,12 @@ mod tests {
         let mars_client = MARSClient::new(Environment::Test, None, MozAdsTelemetryWrapper::noop());
         let ads_client = new_with_mars_client(mars_client);
 
-        let result = ads_client.request_spoc_ads(make_happy_placement_requests(), None, false);
+        let result = ads_client.request_spoc_ads(
+            make_happy_placement_requests(),
+            AdRequestFlags::default(),
+            None,
+            false,
+        );
         assert!(result.is_ok());
         m.assert();
     }
@@ -327,7 +341,12 @@ mod tests {
         let mars_client = MARSClient::new(Environment::Test, None, MozAdsTelemetryWrapper::noop());
         let ads_client = new_with_mars_client(mars_client);
 
-        let result = ads_client.request_tile_ads(make_happy_placement_requests(), None, false);
+        let result = ads_client.request_tile_ads(
+            make_happy_placement_requests(),
+            AdRequestFlags::default(),
+            None,
+            false,
+        );
         assert!(result.is_ok());
         m.assert();
     }
@@ -363,7 +382,12 @@ mod tests {
 
         assert_eq!(client.get_context_id().unwrap(), "custom-context-id-12345");
 
-        let result = client.request_image_ads(make_happy_placement_requests(), None, false);
+        let result = client.request_image_ads(
+            make_happy_placement_requests(),
+            AdRequestFlags::default(),
+            None,
+            false,
+        );
         assert!(result.is_ok());
         m.assert();
     }
@@ -392,7 +416,12 @@ mod tests {
             .create();
 
         let response = ads_client
-            .request_image_ads(make_happy_placement_requests(), None, false)
+            .request_image_ads(
+                make_happy_placement_requests(),
+                AdRequestFlags::default(),
+                None,
+                false,
+            )
             .unwrap();
         let callback_url = response.values().next().unwrap().callbacks.click.clone();
 
@@ -401,7 +430,12 @@ mod tests {
             .create();
 
         ads_client
-            .request_image_ads(make_happy_placement_requests(), None, false)
+            .request_image_ads(
+                make_happy_placement_requests(),
+                AdRequestFlags::default(),
+                None,
+                false,
+            )
             .unwrap();
 
         ads_client.record_click(callback_url, false).unwrap();
@@ -409,6 +443,7 @@ mod tests {
         ads_client
             .request_ads::<AdImage>(
                 make_happy_placement_requests(),
+                AdRequestFlags::default(),
                 Some(CachePolicy::default()),
                 false,
             )